Rename file gettext.php to i12n.php
This is needed to avoid conflicts with php-gettext.
