+++ TODO
@@ -134,7 +134,7 @@ $_REQUEST["json"] wird dabei nicht gesetzt. Eventuell ist das ein Problem der PH
 
 2015-05-12
 
-* gettext.php should find all available translations automatically.
+* i12n.php should find all available translations automatically.
   Currently it must be updated for new translations.
   New translations also need an update of the Linux locales.
   Is this really needed, or is there a better way?
